pyproject.toml
[project]
name = "beancount-importer"
version = "0.1.0"
description = "Various beancount importers"
requires-python = ">=3.11,<4.0"
license = { text = "MIT" }
readme = "README.rst"

authors = [
    { name = "Kevin James", email = "beancount-importer@thekev.in>" },
]

dependencies = [
    "beancount>=3.0.0,<4.0.0",
    "beangulp>=0.2.0,<0.3.0",
    "openpyxl>=3.0.0,<4.0.0",
    "py-pdf-parser>=0.10.0,<0.14.0",
    "python-dateutil>=2.0.0,<3.0.0",
    "titlecase>=2.0.0,<3.0.0",
]

[project.urls]
repository = "https://github.com/TheKevJames/beancount-importer"

[tool.poetry]
requires-poetry = ">=2.0"
package-mode = true
packages = [
    { include = "beancount_importer", format = ["sdist", "wheel"] },
]

[tool.poetry.scripts]
bean-import = "beancount_importer.cli:run"

[[tool.mypy.overrides]]
module = [
    'beancount.core.*',
    'beancount.ingest.*',
    'py_pdf_parser.*',
    'titlecase.*',
]
ignore_missing_imports = true

[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
